Create custom InitBFGS class
This is basically a re-implementation of ROL::lBFGS : ROL::Secant. As
implemented in ROL itself, the lBFGS::applyH() method calls
Secant::applyH0() to get the action of the initial inverse Hessian
estimate. This means that even if we create a subclass of lBFGS, it
won't use any overriden applyH0 method.

Our new InitBFGS class trampolines through PyInitBFGS, which defines
applyH0 as an overridable virtual function. We can then override the
action of the initial inverse Hessian alone, and still retain the rest
of the BFGS/optimisation infrastructure.

Additionally, we can't defer to pybind's constructor binding for
ROL::LineSearchStep, as it won't downcast an InitBFGS (or child
thereof) to a ROL::Secant.