Add applyB to InitBFGS
When using some optimisation steps (such as a trust region), we need
to be able to apply both the inverse Hessian estimate and the forward
Hessian estimate. This adds the applyB method from ROL::lBFGS and
allows overriding of InitBFGS::applyB0 so we can provide our
own (forward) initial Hessian callback.