Add routines for H0/B0 scaling
The original applyH0 and applyB0 routines would simply use the
identity matrix, but if there had been at least one iteration it would
be scaled according to the update vectors. We expose this scaling so
that it can be used by overriding routines of applyH0/B0 by calling
self.scaleH0(Hv) or self.scaleB0(Bv).