Add support for ROL checkpointing
The state contained in ROL can be serialised and loaded through the
cereal library. This requires some fairly minimal modifications to ROL
itself, to get access to private members.

The Vector type (which gets overridden on the Python side) is
serialised by pickling, to be able to de-serialise the correct Python
object. We have external, split save/load functions to call pickle
from the C++ side in the cereal routines. For the load, we need to
step in before cereal allocates memory, because we need to cast an
object returned from Python. This also requires changes on the Python
side to initialise the C++ side of ROL.Vector upon
un-pickling (__init__ doesn't get called from pickle, which is where
we would normally call super().__init__() to do this).

There are separate calls for serialising a ROL Algorithm and Secant,
because the de-serialisation process involves first initialising
these (e.g. initialise the Algorithm with the Secant), then filling in the
member variables.